[UPD] Refactored model of translate
