+++ app/data/common/model/lang/main_window/OMSLDAboutAppModal.interface.ts
