+++ app/data/common/model/lang/main_window/OMSLDLogoMenu.interface.ts
