+++ app/data/common/model/lang/main_window/OMSLDMainMenu.interface.ts
