+++ app/modules/main/header/Logo/Logo.tsx
@@ -7,7 +7,7 @@ import Dropdown from 'antd/lib/dropdown';
 import styles from './Logo.m.scss';
 import { APP_CONSTS } from 'data/utils/AppConsts.enum';
 import { TranslateContext } from 'modules/TranslateContext';
-import IOMSLDLogoMenu from 'data/common/model/lang/OMSLDLogoMenu.interface';
+import IOMSLDLogoMenu from 'data/common/model/lang/main_window/OMSLDLogoMenu.interface';
 
 interface ILogoProps {
   onClickAboutApp?: () => void;
